+++ components/brave_ads/core/internal/account/utility/tokens_feature.h
@@ -13,10 +13,10 @@ namespace brave_ads {
 
 BASE_DECLARE_FEATURE(kAccountTokensFeature);
 
-inline constexpr base::FeatureParam<int> kMinConfirmationTokens{
+inline constexpr base::FeatureParam<size_t> kMinConfirmationTokens{
     &kAccountTokensFeature, "minimum_confirmation_tokens", 20};
 
-inline constexpr base::FeatureParam<int> kMaxConfirmationTokens{
+inline constexpr base::FeatureParam<size_t> kMaxConfirmationTokens{
     &kAccountTokensFeature, "maximum_confirmation_tokens", 50};
 
 }  // namespace brave_ads
