+++ components/brave_ads/core/internal/serving/eligible_ads/priority/priority_util.h
@@ -16,7 +16,7 @@ namespace brave_ads {
 template <typename T>
 void LogNumberOfUntargetedCreativeAdsForBucket(const T& creative_ads,
                                                int priority,
-                                               int bucket) {
+                                               size_t bucket) {
   const size_t count = UntargetedCreativeAdCount(creative_ads);
   if (count > 0) {
     BLOG(3, count << " untargeted ads with a priority of " << priority
@@ -27,7 +27,7 @@ void LogNumberOfUntargetedCreativeAdsForBucket(const T& creative_ads,
 template <typename T>
 void LogNumberOfTargetedCreativeAdsForBucket(const T& creative_ads,
                                              int priority,
-                                             int bucket) {
+                                             size_t bucket) {
   const size_t count = TargetedCreativeAdCount(creative_ads);
   if (count > 0) {
     BLOG(3, count << " targeted ads with a priority of " << priority
