+++ components/brave_ads/core/internal/targeting/contextual/text_classification/text_classification_feature.h
@@ -16,7 +16,7 @@ BASE_DECLARE_FEATURE(kTextClassificationFeature);
 inline constexpr base::FeatureParam<int> kTextClassificationResourceVersion{
     &kTextClassificationFeature, "flatbuffers_resource_version", 1};
 
-inline constexpr base::FeatureParam<int>
+inline constexpr base::FeatureParam<size_t>
     kTextClassificationPageProbabilitiesHistorySize{
         &kTextClassificationFeature, "page_probabilities_history_size", 5};
